Update instructions in the the README for building reference docs
The instructions should have been updated as part of the work on
gh-5267 so that the zip of all of the starter poms is available to
the documentation's build. With that fixed, the current instructions
still fail as the javadoc fails to generated as a result of
org.springframework.boot:spring-boot:test-jar:tests being unavailable.
This can be avoid by simply not trying to build the javadoc, i.e.
by running generate-resources rather than install.

Closes gh-5633