Try to fix the CI build
The order of the modules declared in pom.xml should not make any
difference given the other relationships that exist between projects,
however the CI build is ordering the projects incorrectly.

Try changing the order of the modules to see if the CI build can be
coaxed into ordering them correctly.