Add Jackson Kotlin module to JSON starter
This commit adds the `jackson-module-kotlin` dependency to the JSON
starter. As the other modules selected there, the goal is to provide
additional Jackson support to Spring Boot projects.

In this case, the Kotlin language is supported by many Spring projects
and this module is essential for (de)serialization with Jackson.
Note that the module has a transitive dependency on `kotlin-reflect`,
but this dependency should be brought by the application itself.

Spring Framework will configure the Kotlin Jackson module *only if* the
Kotlin sdk is present.

Closes gh-9803