Stop included profiles overriding active profiles
Update `ConfigFileApplicationListener` so that `spring.profiles.include`
properties do not override higher priority active profiles.

This commit also changes when profiles get added to the environment.
Profiles are now added to the environment at the time of profile
processing so that they get logged in the order that they are processed.

Closes gh-13151